# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "L5 Swagger API",
        "description": "L5 Swagger API description",
        "contact": {
            "email": "info@wirelessranking.com"
        },
        "version": "1.0.0"
    },
    "host": "ranking.dev.sapient.pro",
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
        "/countries": {
            "get": {
                "tags": [
                    "General"
                ],
                "summary": "Get list of countries",
                "description": "Get list of countries",
                "operationId": "countries",
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
...
        "/reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Get reviews for all the operators",
                "description": "Get reviews for all the operators",
                "operationId": "reviews",
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:projects",
                            "read:projects"
                        ]
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "passport": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "/oauth/token",
            "scopes": {}
        }
    }
}